% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6class_composite_milk_record.R
\docType{data}
\name{R6ClassCompositeMilkRecord}
\alias{R6ClassCompositeMilkRecord}
\title{R6Class for Composite Milk Record}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6ClassCompositeMilkRecord
}
\description{
The R6Class \code{R6ClassCompositeMilkRecord} represents objects
for modelling composite Milk records. A composite milk record
is defined to consist of the following components
\itemize{
  \item milkYield
  \item fatPercent
  \item proteinPercent
  \item sccValue
  \item ureaContent
  \item lactoseContent
}
Each of those components is modelled as objects of \code{R6ClassGenericRecord}.
}
\keyword{datasets}

