% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6class_generic_record.R
\docType{data}
\name{R6ClassGenericRecord}
\alias{R6ClassGenericRecord}
\title{R6 Class for Generic Data Records}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6ClassGenericRecord$new()
}
\value{
Object of \code{\link{R6Class}} R6ClassGenericRecord
}
\description{
The R6 class \code{R6ClassGenericRecord} represents
objects of generic data records. These generic objects
can be used derive more specific data records such as
milk yield records, fat percentage records, etc.
}
\section{Fields}{

\describe{
\item{\code{recordType}}{name of record type}

\item{\code{recordDate}}{date of recording}

\item{\code{recordValue}}{value of the record}

\item{\code{recordUnit}}{unit associated to record value}
}}
\section{Public methods}{

\describe{
  \item{\code{new()}}{Instantiation of an R6ClassGenericRecord object}
  \item{\code{sampleUnifValue(minValue = 0, maxValue = 1)}}{Sample a value from a uniform
                distribution between \code{minValue} and \code{maxValue}}
  \item{\code{toTsvString()}}{Conversion of a generic record to a string using TABs as field
                separators. The order in which the fields are pasted together is fixed to
                Type Date Value Unit.}
}
}
\examples{
r6objGenRec <-  R6ClassGenericRecord$new()
r6objGenRec$setRecordType(recordType = "MilkYield")
r6objGenRec$setRecordValue(recordValue = 32.6)
r6objGenRec$setRecordUnit(recordUnit = "kg")
\dontrun{
r6objGenRec$toTsvString()
}
}
\keyword{datasets}

